cloud_computing/function/requirements.txt
# Function dependencies, for example:
# package>=version

tensorflow==2.0.0b0
google-cloud-storage==1.16.1
Pillow==6.0.

